+++ src/utilities/http/index.ts
@@ -1,6 +1,4 @@
-export * from "./deploy";
 export * from "./instruments";
 export * from "./auditLogs";
 export * from "./upload";
 export * from "./toStartDate";
-export * from "./removeInstrument";
