+++ src/utilities/processes/index.ts
@@ -1,0 +1,2 @@
+export * from "./deployInstrument";
+export * from "./removeInstrument";
